Restructure, add OutputId() and utf8-hex conversion functions
