+++ sdk/src/client/bindings/python/iota_client/__init__.py
@@ -1,3 +1,9 @@
 from .iota_client import *
 from .client import *
 from .secret_manager import *
+from .prefix_hex import *
+from .types.address import *
+from .types.common import *
+from .types.feature import *
+from .types.output_id import *
+from .types.unlock_condition import *
